+++ repositories/settings/core.go
@@ -1,16 +1,16 @@
 package settings
 
 type core struct {
-	fovY   float32
+	fovY   float64
 	width  int32
 	height int32
 }
 
-func (c *core) setFOV(degY float32) {
+func (c *core) setFOV(degY float64) {
 	c.fovY = degY
 }
 
-func (c *core) getFOV() float32 {
+func (c *core) getFOV() float64 {
 	return c.fovY
 }
